Write a MainView class and DriveStatus for listing
We now present the user with a single view that lists all known backup
targets and their backup status. We use contextual buttons to allow for
the removal of backup targets or the restoring from them.

This is enabled by the new BackupTargetManager class. We have made all
of the older screens BackupMessages which are displayed above this
backup target listing.