Use signals to communicate BackupStatus -> MainView
Before we would update the BackupTarget list in place then ask MainView
to do a full redraw. This was exteremly inefficient and potentially
error-prone. Now backup status sends out signals indicating the status
of a target as it finds it. Application now wires up those signals to
pass the appropriate status for a target to MainView which updates that
single entry