Add NEEDS_BACKUP status use new icon for NEEDS_BACKUP vs ERROR
Closes #3

The process-stop icon now indicates a backup error. We use a less scary
dialog-warning icon to indicate the backup is out of date. Backup error
indicates that user intervention is required. We additionally track last
status for a DriveStatus so that if something tries to set it to any
state but safe from ERROR we refuse to update to a "false" status.