No longer test for differing files
There is a bug with storing backup targets last_backup_sources in JSON.
I believe this is a bug in GLib.JSON but I haven't been able to track it
down. As it stands if you have multiple drives whenever you backup to
the one of them one or both of the drives (it's non-deterministic) will
serialize their last_backup_sources as nothing.

For now simple "older than 24 hours" or never had a backup checks are
fine.